% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise_and_insults.R
\name{praise_function}
\alias{praise_function}
\title{Geussing who staff is!}
\usage{
praise_function(purpose = NULL, lines_of_code = NULL)
}
\arguments{
\item{purpose}{either Stats, Data or Text Mining}

\item{lines_of_code}{integar variable}
}
\value{

}
\description{
when you run the function with a purpose and lines of code it will guess who the staff member is between Zoe, Chris, Milan and Andreas
}
\examples{
praise_function(purpose = "data", lines_of_code = 200)
 "Zoe is outstanding"



}
